pyproject.toml
[tool.poetry]
name = "wordle-parser"
version = "0.1.0"
description = ""
authors = ["Tessmore <fabientesselaar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
openpyxl = "^3.1.5"
python-dateutil = "^2.9.0.post0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
